+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [next]
 
 - chore(TS): control generic object type [#9770](https://github.com/fabricjs/fabric.js/pull/9770)
+- refactor(): `_findTargetCorner` is now called `findControl` and returns the key and the control and the coordinates [#9668](https://github.com/fabricjs/fabric.js/pull/9668)
 - feat(LayoutManager): Handle the case of activeSelection with objects inside different groups [#9651](https://github.com/fabricjs/fabric.js/pull/9651)
 
 ## [6.0.0-beta20]